pyproject.toml
[tool.poetry]
name = "drug-ner"
version = "0.1.0"
description = ""
authors = ["George Kontonatsios <g.kontonatsios@thesignalgroup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
psycopg2 = "^2.9.9"
loguru = "^0.7.2"
requests = "^2.32.3"
openai = "^1.32.0"
tqdm = "^4.66.4"
torchvision = "*"
torchaudio = "*"
transformers = {extras = ["torch"], version = "^4.41.2"}
datasets = "^2.19.2"
scikit-learn = "^1.5.0"
sentencepiece = "^0.2.0"
streamlit = "^1.35.0"
st-annotated-text = "^4.0.1"
pydantic-settings = "^2.3.1"
 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
